backend/pyproject.toml
# backend/pyproject.toml
# Urban Intelligence Framework v2.0.0
# Backend Python project configuration managed with UV

[project]
name = "urban-intelligence-backend"
version = "2.0.0"
description = "Urban Intelligence Framework — ML backend for Airbnb price prediction"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Francisco Javier Mercader Martínez", email = "fcojavier.mercader04@gmail.com" },
]
keywords = [
  "airbnb",
  "machine-learning",
  "polars",
  "duckdb",
  "mlflow",
  "fastapi",
  "xgboost",
]

dependencies = [
  # ── Core data ─────────────────────────────────────────────────────────────
  "polars>=1.0.0",
  "duckdb>=1.0.0",
  "numpy>=1.26.0",
  "pyarrow>=17.0.0",
  "pandas>=2.2.0",

  # ── Machine learning ──────────────────────────────────────────────────────
  "mlflow>=2.15.0",
  "xgboost>=2.1.0",
  "lightgbm>=4.4.0",
  "catboost>=1.2.5",
  "optuna>=3.6.0",
  "scikit-learn>=1.5.0",
  "scipy>=1.13.0",
  "statsmodels>=0.14.0",

  # ── NLP / Transformers ────────────────────────────────────────────────────
  "transformers>=4.42.0",
  "torch>=2.3.0",
  "tokenizers>=0.19.0",
  "sentencepiece>=0.2.0",

  # ── Computer vision ───────────────────────────────────────────────────────
  "Pillow>=10.4.0",
  "torchvision>=0.18.0",

  # ── Web API ───────────────────────────────────────────────────────────────
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "strawberry-graphql[fastapi]>=0.232.0",
  "websockets>=12.0",
  "httpx>=0.27.0",

  # ── Config & validation ───────────────────────────────────────────────────
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",

  # ── Caching ───────────────────────────────────────────────────────────────
  "redis>=5.0.0",
  "hiredis>=2.3.0",

  # ── Scraping ──────────────────────────────────────────────────────────────
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",

  # ── Monitoring & observability ────────────────────────────────────────────
  "structlog>=24.1.0",
  "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
  "pre-commit>=3.7.0",
  "bandit>=1.7.0",
]

[project.scripts]
urban-etl = "scripts.run_etl:main"
urban-train = "scripts.run_training:main"
urban-api = "api.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "api", "scripts"]

[tool.ruff]
target-version = "py311"
line-length = 79

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["N803", "B008", "E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]
